+++ docs/notebooks/sl.txt
@@ -1,0 +1,802 @@
+0.00	0
+0.003	0
+0.006	-10.07018
+0.009	-37.63985
+0.012	-80.35509
+0.015	-114.70685
+0.018	-129.57405
+0.021	-125.95758
+0.024	-119.71684
+0.027	-113.09509
+0.03	-91.62696
+0.033	-86.99771
+0.036	-78.35769
+0.039	-78.12284
+0.042	-72.19138
+0.045	-69.08671
+0.048	-64.55599
+0.051	-59.60964
+0.054	-54.68657
+0.057	-69.17602
+0.06	-83.73558
+0.063	-85.16648
+0.066	-68.82416
+0.069	-64.8212
+0.072	-54.02043
+0.075	-40.54444
+0.078	-36.14664
+0.081	-22.36629
+0.084	-40.88312
+0.087	-43.818
+0.09	-42.55527
+0.093	-41.08478
+0.096	-38.03709
+0.099	-26.40847
+0.102	-34.71745
+0.105	-43.05675
+0.108	-39.60518
+0.111	-30.02263
+0.114	-20.85337
+0.117	-10.48935
+0.12	-15.31545
+0.123	-11.06967
+0.126	-46.45587
+0.129	-98.17449
+0.132	-137.28666
+0.135	-137.47435
+0.138	-137.66094
+0.141	-119.78474
+0.144	-115.13597
+0.147	-104.74175
+0.15	-104.42788
+0.153	-109.53298
+0.156	-114.61205
+0.159	-108.16204
+0.162	-97.09253
+0.165	-86.7531
+0.168	-76.61874
+0.171	-68.15603
+0.174	-67.65087
+0.177	-72.56844
+0.18	-71.60184
+0.183	-64.39172
+0.186	-62.21436
+0.189	-41.53983
+0.192	-46.36795
+0.195	-28.30365
+0.198	-16.22634
+0.201	-23.42822
+0.204	-26.953
+0.207	-28.49322
+0.21	-26.26003
+0.213	-28.26454
+0.216	-32.01557
+0.219	-61.75538
+0.222	-71.0083
+0.225	-65.84401
+0.228	-55.79992
+0.231	-49.6035
+0.234	-41.62994
+0.237	-19.5832
+0.24	-49.30356
+0.243	-79.19226
+0.246	-93.61458
+0.249	-84.56446
+0.252	-77.24887
+0.255	-78.67626
+0.258	-73.4265
+0.261	-69.31015
+0.264	-74.97886
+0.267	-75.64383
+0.27	-80.21326
+0.273	-70.23462
+0.276	-69.78598
+0.279	-60.81059
+0.282	-49.73811
+0.285	-41.30203
+0.288	-42.14879
+0.291	-51.89427
+0.294	-46.03414
+0.297	-65.33404
+0.3	-57.74642
+0.303	-46.81681
+0.306	-51.67572
+0.309	-33.86392
+0.312	-31.18764
+0.315	-32.00314
+0.318	-26.65905
+0.321	-37.07859
+0.324	-35.31693
+0.327	-28.99564
+0.33	-42.81685
+0.333	-59.34329
+0.336	-92.88223
+0.339	-101.36891
+0.342	-97.8504
+0.345	-98.32066
+0.348	-93.03739
+0.351	-76.30229
+0.354	-72.85584
+0.357	-65.61489
+0.36	-66.85444
+0.363	-46.57714
+0.366	-68.06831
+0.369	-47.85425
+0.372	-21.9749
+0.375	-26.93116
+0.378	-41.91219
+0.381	-22.4419
+0.384	-28.13371
+0.387	-6.13882
+0.39	-23.9916
+0.393	-13.06959
+0.396	-14.45414
+0.399	-24.82528
+0.402	-4.61552
+0.405	-6.4947
+0.408	-12.09021
+0.411	-36.65932
+0.414	-36.89559
+0.417	-49.24717
+0.42	-82.97271
+0.423	-95.73724
+0.426	-128.56049
+0.429	-141.61142
+0.432	-125.66042
+0.435	-114.75125
+0.438	-111.4508
+0.441	-100.51396
+0.444	-102.65188
+0.447	-85.20376
+0.45	-75.4119
+0.453	-90.80341
+0.456	-74.23337
+0.459	-54.54862
+0.462	-71.54363
+0.465	-62.96249
+0.468	-28.74713
+0.471	-37.99759
+0.474	-29.49675
+0.477	-26.10749
+0.48	-21.46605
+0.483	-17.89569
+0.486	-11.29516
+0.489	-34.14434
+0.492	-24.78668
+0.495	-45.95628
+0.498	-42.61654
+0.501	-48.33655
+0.504	-52.63022
+0.507	-68.0878
+0.51	-71.52008
+0.513	-75.76562
+0.516	-57.81472
+0.519	-57.2478
+0.522	-58.169
+0.525	-62.14236
+0.528	-62.96463
+0.531	-61.17955
+0.534	-51.00394
+0.537	-69.49645
+0.54	-31.66296
+0.543	-45.32759
+0.546	-39.17455
+0.549	-36.95302
+0.552	-28.20462
+0.555	-45.68023
+0.558	-26.41236
+0.561	-30.16165
+0.564	-19.64488
+0.567	-16.23789
+0.57	-12.50457
+0.573	-41.61278
+0.576	-17.74273
+0.579	-48.86405
+0.582	-36.65589
+0.585	-51.26317
+0.588	-50.11973
+0.591	-42.91791
+0.594	-41.30568
+0.597	-34.9636
+0.6	-64.51231
+0.603	-61.72329
+0.606	-36.14102
+0.609	-61.83831
+0.612	-47.70505
+0.615	-30.68884
+0.618	-44.53392
+0.621	-90.80587
+0.624	-119.16068
+0.627	-143.50919
+0.63	-126.08501
+0.633	-102.53048
+0.636	-109.3652
+0.639	-97.01719
+0.642	-78.60763
+0.645	-79.87385
+0.648	-79.96701
+0.651	-76.35062
+0.654	-75.94489
+0.657	-76.2894
+0.66	-75.51278
+0.663	-81.93515
+0.666	-91.20741
+0.669	-65.10661
+0.672	-70.78136
+0.675	-57.13214
+0.678	-47.5153
+0.681	-64.60211
+0.684	-48.88776
+0.687	-12.38549
+0.69	-35.64091
+0.693	-42.61883
+0.696	-35.46524
+0.699	-36.58035
+0.702	-50.88342
+0.705	-50.68347
+0.708	-78.03455
+0.711	-88.38896
+0.714	-123.46187
+0.717	-83.72826
+0.72	-96.55417
+0.723	-52.93273
+0.726	-66.30332
+0.729	-54.71593
+0.732	-19.59868
+0.735	-74.15515
+0.738	-45.04499
+0.741	-65.8351
+0.744	-77.53277
+0.747	-90.52657
+0.75	-88.48062
+0.753	-66.1249
+0.756	-27.83762
+0.759	-35.62981
+0.762	-66.09392
+0.765	-59.03224
+0.768	-38.95528
+0.771	-21.21365
+0.774	0.80881
+0.777	-29.75643
+0.78	-30.47398
+0.783	-53.77896
+0.786	-49.86752
+0.789	-79.01945
+0.792	-105.95225
+0.795	-77.93918
+0.798	-90.17097
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+
